Moved locust.cache into locust.util.cache
